+++ Resources/views/frontend/account/sidebar.tpl
@@ -22,6 +22,11 @@
             </a>
         </li>
 
+        {if isset($atsdFirstName) }
+            {$userInfo['firstname'] = $atsdFirstName}
+        {/if}
+
+
     {/if}
 
 {/block}
